use the top 8-bits more accurately for void-extent blocks in sRGB mode
Section 2.5. Table 4, mentions that an 8-bit resolution is allowed for sRGB
mode. After masking the bottom 8 bits in the pixel channels of void-extent
blocks in sRGB mode, the resulting value is treated as a UNORM16. This causes
all input values above 128 (in the upper 8 bits) to return an output that
is 1 less than expected. For example, 129 in the top 8 bits returns 128.
This patch uses the top 8 bits to generate a float that will give the
expected output when writing out to disk.