Add command line functional test suite (#103)
This PR adds the beginnings of a command line functional test suite, ensuring that command line options are correctly wired up and processed. For this first PR not every option is tested, but this suite includes positive tests and negative tests for all of the common operations. See the testing documentation page for more details on how to run these tests.

Some negative tests are currently marked as skipped pending bug fix (notably, the code is intolerant of missing input files, or output directories), so the test suite does currently pass on both Windows and Linux (macOS should pass, but is not tested).

In addition this commit increases the astcenc stack size on Windows to 4MB. We have had persistent problem with stack space on Windows builds, which only have 1MB stacks by default compared to Linux which reserves 8MB. Something in a recent VS2019 upgrade make the build use more stack space which just pushed us over the 1MB limit, so bumped these up to 4MB which should be more than enough.