Implement VLA compute_error_of_weight_set
This implementation vectorizes the implementation of this function
using the vector-length agnostic SIMD support. To make this
vectorizable we've switched some structures from AOS to SOA, which
helps this function, but makes the code slow down elsewhere.

Loosely this optimization is 4-6% faster for -thorough, 1-3%
faster for -medium, neutral for -fast, and 2-3% slower for -fastest.