Optimize compute_error_squared_rgb/a()
This implementation alters the algorithm and vectorizes the
resulting function for this function.

The algorithm change will have a minor image quality impact; we
use a faster approximation for selecting which component to assign
to a second plane when choosing a partition. There is a very minor
quality hit on -medium (under < 0.01dB), and a slightly larger one
on -thorough (under < -0.05dB). The significant (>10%) performance
gain is worth  the quality loss (i.e. users are more likely to
use -thorough if it's faster, and the overall gain vs medium is
still net positive).