Make codec invariant across None/SSE/AVX2 builds (#224)
This patch makes the none (no SIMD), SSE, and AVX2 builds invariant with respect to each other. Prior to this only the SSE2 and 4.1 builds were invariant. This is a  fix for #223.

Four classes of change were needed to avoid reassociation rounding differences:

* Horizontal sums in a vector must match the odd-even reduce used by the SIMD cases, even for non-SIMD code, to ensure consistent reduction sums.
* All vector body accumulators that use lane-wise accumulation must use the same accumulator width, irrespective of instruction set, to ensure consistent lane-wise partial sums. For this reason the 1-wide VLA fallback has been removed, as we decided to use 4-wide accumulators. 
* AVX2 vector body accumulators must split into two 4-wide operations, and sequentially adding these into the running sum, to ensure consistent accumulator partial sums.  
* Vector loop bodies that use a loop tail for accumulation must execute the same number of times, irrespective of instruction set, to ensure that the loop tail runs the same number of times. This is because a the vector body will do odd+even partial sums (reduction across the vector), whereas the loop tail will do linear partial sums (iterating though the remaining tail). 

For this patch I've implemented the last bullet simply by rounding down all vector bodies that use accumulators and loop tails to a multiple of 8. This will reduce the number of vectorized loop iterations for 4-wide hardware in some cases, e.g. for an iteration count of 13, the left over is 5, which would have run one more vectorized path. Measured impact is small and only hits some block sizes, although 6x6 is impacted by ~2% on -thorough. This will be optimized in a follow-on patch.