Remove weight infill texel count from decimation error
The effect of this change is to very slightly reduce PSNR, but it
should improve consistency and reduce block artifacts as weights
for block-edge texels are no longer getting treated as less
important than weights for mid-block texels, just because they
influence fewer texels.

Gains 2% perforamnce across the board for all quality levels.