Limit weight matching to 32 weights
We were matching 33 weights for QUANT_32 to avoid the double-size
hole in the middle of the weight space for 32. However this forces
a whole vector iteration with only a single lane active.

Removing this is marginally lower quality (0.005 dB), but 2-3%
faster, so definitely a change worth making. With this change
we can also reduce ANGULAR_STEPS to 32.