Use sized formats in decompressed KTX output files (#354)
KTX1 requires gl_internal_format to be a sized format. To meet this requirement 1 component luminance textures are now using GL_R and  R(8/16), and 2 component luminance+alpha textures are using  GL_RG + GL_RG(8/16). 

This is a semantic change from the old unsized GL_LUMINANCE and LUMINANCE_ALPHA, but consistent with the future direction of KTX2 (although we cannot specify a KTX2-style KTXswizzle here, as the namespace is reserved in KTX1).