Enable building x86_64h slices on macOS
This additional x86_64 variant target only exists on macOS and represents a Haswell-or-later CPU—i.e., with AVX2 available.  It can be used as an additional slice of a universal binary alongside x86_64 and arm64, enabling a single universal binary to get maximal performance on all supported Macs, since the system will automatically use the x86_64h slice if it's available and the CPU supports it.