Cheaper compute_lowest_and_highest_weight
The actual weights array is constant, so we can compute the
min/max weights in a single pass ahead of time. The remapping
we apply to the weights depending on rcp_stepsize and offset
is monotonic; therefore we can figure out minidx and maxidx in
advance from the min/max weights we determined previously.

This in turn means that we don't need to do the "reset" handling
in the loop where we zero the accumulator whenever the current
min/max changes. The amount of code in the inner loop nest is
considerably reduced as a result, and the extra preprocessing
outside that loop is generally less than the work it replaces.

The exact impact depends on the block size and max_angular_steps.
Testing with RGB textures and -thorough quality, I observe
approximately 0.7% coding time reduction with 4x4 blocks; 6x6
and larger see a 1.2% reduction.

At -fast quality, there is not much gain, but neither is it
any slower.

In my tests, this change does not alter encoding results (nor
should it); the output is bitwise identical to before.

Signed-off-by: Fabian Giesen <fabian.giesen@epicgames.com>