Use relative sum of squares error for HDR textures
HDR texture values are stored logarithmically.

Using absolute sum of squares on the logarithmic values causes the
compressor to spend too much effort preserving imperceptible shifts in
dark channel values at the expense of bright values in the same block.
This performs poorly in blocks with sharp luminance changes (dark
texels) and in blocks with saturated color values (dark channels in
bright pixels).

Using absolute sum of squares on linearized HDR values avoids the
compressor fixating on dark values, but instead causes the compressor to
spend too much effort preserving bright values. This is because the
errors in the bright channels can be orders of magnitude bigger than
the errors in the dark channels, and dark values can end up quantizing
close to black.

Using relative sum of square on the logarithmic values, proposed by Ryg
in the blog below, encourages the compressor to find a balance of
relative error across the whole block, favoring neither light nor
dark channels.

https://fgiesen.wordpress.com/2024/11/14/mrsse/