Align with KDFS 1.4.0 rounding update (#548)
Updates the codec rounding for sRGB to align with the KDFS 1.4.0
specification update, which reverts the alpha channel endpoint 
expansion behavior to match an earlier version of the specification.

This changes the alpha channel to use the same method 
((x << 8) | 0x80) as the RGB channels. The previous behavior for 
alpha was doing ((x << 8) | x). The impact of this change is 
differences in the LSB for sRGB images.

Fixes #547