Use idiomatic CLZ implementation (#582)
Replaces the previous implementation of clz32 in the reference softfloat library, which had multiple paths and compiler-specific intrinsics or inline assembly for some architectures, with a simpler idiomatic pure C implementation.

Modern compilers recognize this implementation and will optimize into a native CLZ instruction on most architectures. Tested on Clang 18 and GCC 16 for AArch64 and x86-64.